+++ concordium-consensus/src/Concordium/MultiVersion.hs
@@ -23,9 +23,9 @@
 module Concordium.MultiVersion where
 
 import Control.Concurrent
-import Control.Exception
+import Control.Exception hiding (handle)
 import Control.Monad
-import Control.Monad.Catch (MonadCatch, MonadMask, MonadThrow (throwM))
+import Control.Monad.Catch (MonadCatch, MonadMask, MonadThrow (throwM), handle)
 import Control.Monad.Reader
 import qualified Control.Monad.State.Strict as State
 import Data.ByteString (ByteString)
@@ -34,6 +34,7 @@ import Data.IORef
 import Data.Serialize
 import qualified Data.Text as Text
 import Data.Time
+import Data.Typeable
 import qualified Data.Vector as Vec
 import System.FilePath
 
@@ -500,6 +501,26 @@ data CatchUpStatusBufferState
     | -- | We should not send any more catch-up status messages.
       BufferShutdown
 
+-- | A representation of whether the global write lock has been poisoned due to an exception in
+--  a thread that held the lock.
+data GlobalStateStatus
+    = -- | The global state is in a consistent state.
+      GSOk
+    | -- | A thread has thrown an exception while processing the global state, indicating an
+      --  unrecoverable error.
+      GSError
+    deriving (Eq)
+
+-- | An exception indicating that the global state lock was poisoned by an error in another thread.
+--  This indicates that some invariants of the global state may have been violated, and is not
+--  recoverable.
+data GlobalLockPoisonError = GlobalLockPoisonError
+    deriving (Eq, Show, Typeable)
+
+instance Exception GlobalLockPoisonError where
+    displayException _ =
+        "The global state lock was poisoned by an unrecoverable error in another thread."
+
 -- | The context for managing multi-version consensus.
 data MultiVersionRunner finconf = MultiVersionRunner
     { -- | Base configuration.
@@ -515,7 +536,7 @@ data MultiVersionRunner finconf = MultiVersionRunner
       --  the write lock is held.
       mvVersions :: !(IORef (Vec.Vector (EVersionedConfiguration finconf))),
       -- | Global write lock.
-      mvWriteLock :: !(MVar ()),
+      mvWriteLock :: !(MVar GlobalStateStatus),
       -- | Flag to stop importing blocks. When importing blocks from a file is in progress,
       --  setting this flag to True will cause the import to stop.
       mvShouldStopImportingBlocks :: !(IORef Bool),
@@ -546,6 +567,25 @@ instance MonadLogger (MVR finconf) where
     logEvent src lvl msg = MVR $ \mvr -> mvLog mvr src lvl msg
     {-# INLINE logEvent #-}
 
+-- | Catch and log exceptions in the 'MVR' monad.
+--  Returns a specified value in the event of an exception.
+handleMVRExceptionsWith ::
+    -- | Value to return in the event of an exception.
+    a ->
+    -- | Action to run
+    MVR finconf a ->
+    MVR finconf a
+handleMVRExceptionsWith failRes = handle handler
+  where
+    handler (e :: SomeException) = do
+        logEvent Runner LLError $ "Unrecoverable exception: " <> displayException e
+        return failRes
+
+-- | Catch and log exceptions in the 'MVR' monad.
+--  Returns 'Skov.ResultConsensusFailure' in the event of an exception.
+handleMVRExceptions :: MVR finconf Skov.UpdateResult -> MVR finconf Skov.UpdateResult
+handleMVRExceptions = handleMVRExceptionsWith Skov.ResultConsensusFailure
+
 -- | Perform an action while holding the global state write lock.
 --  If the action throws an exception, this ensures that the lock is
 --  released.
@@ -561,14 +601,23 @@ withWriteLockIO :: MultiVersionRunner finconf -> IO a -> IO a
 withWriteLockIO MultiVersionRunner{..} a =
     bracketOnError
         (takeMVar mvWriteLock)
-        ( \() ->
-            tryPutMVar mvWriteLock ()
-                >> mvLog Runner LLWarning "Released global state lock following error."
+        ( \_ -> do
+            poisoned <- tryPutMVar mvWriteLock GSError
+            when poisoned $
+                mvLog
+                    Runner
+                    LLWarning
+                    "An error occurred while holding the global state lock.\
+                    \ This will be propagated to other threads."
         )
-        $ \_ -> do
-            res <- a
-            putMVar mvWriteLock ()
-            return res
+        $ \case
+            GSOk -> do
+                res <- a
+                putMVar mvWriteLock GSOk
+                return res
+            GSError -> do
+                putMVar mvWriteLock GSError
+                throwIO GlobalLockPoisonError
 
 -- | Perform an action while holding the global state write lock. Optionally, when the action
 --  completes, a thread is forked to perform a follow-up action before releasing the lock.
@@ -585,17 +634,26 @@ withWriteLockMaybeFork action followup = MVR $ \mvr ->
 withWriteLockMaybeForkIO :: MultiVersionRunner finconf -> IO (a, Maybe b) -> (b -> IO ()) -> IO a
 {-# INLINE withWriteLockMaybeForkIO #-}
 withWriteLockMaybeForkIO MultiVersionRunner{..} action followup = mask $ \unmask -> do
-    () <- takeMVar mvWriteLock
-    (res, mContinue) <- unmask action `onException` tryPutMVar mvWriteLock ()
+    gsStatus <- takeMVar mvWriteLock
+    when (gsStatus == GSError) $ do
+        putMVar mvWriteLock GSError
+        throwIO GlobalLockPoisonError
+    (res, mContinue) <- unmask action `onException` tryPutMVar mvWriteLock GSError
     case mContinue of
         Just continueArg -> do
-            let release = putMVar mvWriteLock ()
+            let release (Left exception) = do
+                    mvLog Runner LLError $
+                        "An unrecovarable error occurred in a worker thread\
+                        \ while holding the global state lock: "
+                            ++ show exception
+                    putMVar mvWriteLock GSError
+                release (Right _) = putMVar mvWriteLock GSOk
             -- forkIO is guaranteed to be uninterruptible, so we can be sure that an async exception
             -- won't prevent the lock being released. Also note that the masking state of the thread
             -- is inherited, so we unmask when running the follow-up.
-            void $ forkIO (unmask (try @SomeException (followup continueArg)) >> release)
+            void $ forkIO (unmask (try @SomeException (followup continueArg)) >>= release)
         Nothing -> do
-            putMVar mvWriteLock ()
+            putMVar mvWriteLock GSOk
     return res
 
 -- | Lift a 'LogIO' action into the 'MVR' monad.
@@ -1137,7 +1195,7 @@ makeMultiVersionRunner
         mvTransactionPurgingThread <- newEmptyMVar
         let mvr = MultiVersionRunner{..}
         runMVR (startupSkov genesis) mvr
-        putMVar mvWriteLock ()
+        putMVar mvWriteLock GSOk
         startTransactionPurgingThread mvr
         return mvr
 
@@ -1467,12 +1525,14 @@ shutdownMultiVersionRunner MultiVersionRunner{..} = mask_ $ do
     -- Kill the transaction purging thread, if any.
     tryTakeMVar mvTransactionPurgingThread >>= mapM_ killThread
     -- Acquire the write lock. This prevents further updates, as they will block.
-    takeMVar mvWriteLock
-    versions <- readIORef mvVersions
-    -- Shut down the consensus databases.
-    runLoggerT (forM_ versions evcShutdown) mvLog
-    -- Shut down the global account map.
-    LMDBAccountMap.closeDatabase (globalAccountMap (mvcStateConfig mvConfiguration))
+    takeMVar mvWriteLock >>= \case
+        GSOk -> do
+            versions <- readIORef mvVersions
+            -- Shut down the consensus databases.
+            runLoggerT (forM_ versions evcShutdown) mvLog
+            -- Shut down the global account map.
+            LMDBAccountMap.closeDatabase (globalAccountMap (mvcStateConfig mvConfiguration))
+        GSError -> throwIO GlobalLockPoisonError
 
 -- | Lift a version-0 consensus skov action to the 'MVR' monad, running it on a
 --  particular 'VersionedConfigurationV0'. Note that this does not
@@ -1678,58 +1738,67 @@ receiveBlock ::
     GenesisIndex ->
     ByteString ->
     MVR finconf (Skov.UpdateResult, Maybe ExecuteBlock)
-receiveBlock gi blockBS = withLatestExpectedVersion gi $ \case
-    (EVersionedConfigurationV0 (vc :: VersionedConfigurationV0 finconf pv)) -> do
-        MVR $ \mvr -> do
-            now <- currentTime
-            case deserializeExactVersionedPendingBlock (protocolVersion @pv) blockBS now of
-                Left err -> do
-                    mvLog mvr Runner LLDebug err
-                    return (Skov.ResultSerializationFail, Nothing)
-                Right block -> do
-                    (updateResult, mVerifiedPendingBlock) <- runMVR (runSkovV0Transaction vc (Skov.receiveBlock block)) mvr
-                    case mVerifiedPendingBlock of
-                        Nothing -> return (updateResult, Nothing)
-                        Just verifiedPendingBlock -> do
-                            let exec = do
-                                    runSkovV0Transaction vc (Skov.executeBlock verifiedPendingBlock)
-                            let cont = ExecuteBlock $ runMVR exec mvr
-                            return (updateResult, Just cont)
-    (EVersionedConfigurationV1 (vc :: VersionedConfigurationV1 finconf pv)) -> do
-        MVR $ \mvr -> do
-            now <- currentTime
-            case SkovV1.deserializeExactVersionedPendingBlock @pv blockBS now of
-                Left err -> do
-                    mvLog mvr Runner LLDebug err
-                    return (Skov.ResultSerializationFail, Nothing)
-                Right block -> do
-                    blockResult <- runMVR (runSkovV1Transaction vc (SkovV1.uponReceivingBlock block)) mvr
-                    case blockResult of
-                        SkovV1.BlockResultSuccess vb -> do
-                            let exec = do
-                                    runSkovV1Transaction vc (SkovV1.executeBlock vb)
-                                    return Skov.ResultSuccess
-                            let cont = ExecuteBlock $ runMVR exec mvr
-                            return (Skov.ResultSuccess, Just cont)
-                        SkovV1.BlockResultDoubleSign vb -> do
-                            runMVR (runSkovV1Transaction vc (SkovV1.executeBlock vb)) mvr
-                            return (Skov.ResultDoubleSign, Nothing)
-                        SkovV1.BlockResultInvalid -> return (Skov.ResultInvalid, Nothing)
-                        SkovV1.BlockResultStale -> return (Skov.ResultStale, Nothing)
-                        SkovV1.BlockResultPending -> return (Skov.ResultPendingBlock, Nothing)
-                        SkovV1.BlockResultEarly -> return (Skov.ResultEarlyBlock, Nothing)
-                        SkovV1.BlockResultDuplicate -> return (Skov.ResultDuplicate, Nothing)
-                        SkovV1.BlockResultConsensusShutdown -> return (Skov.ResultConsensusShutDown, Nothing)
+receiveBlock gi blockBS = handleMVRExceptionsWith (Skov.ResultConsensusFailure, Nothing) $
+    withLatestExpectedVersion gi $ \case
+        (EVersionedConfigurationV0 (vc :: VersionedConfigurationV0 finconf pv)) -> do
+            MVR $ \mvr -> do
+                now <- currentTime
+                case deserializeExactVersionedPendingBlock (protocolVersion @pv) blockBS now of
+                    Left err -> do
+                        mvLog mvr Runner LLDebug err
+                        return (Skov.ResultSerializationFail, Nothing)
+                    Right block -> do
+                        (updateResult, mVerifiedPendingBlock) <-
+                            runMVR
+                                (runSkovV0Transaction vc (Skov.receiveBlock block))
+                                mvr
+                        case mVerifiedPendingBlock of
+                            Nothing -> return (updateResult, Nothing)
+                            Just verifiedPendingBlock -> do
+                                let exec = do
+                                        runSkovV0Transaction vc $
+                                            Skov.executeBlock verifiedPendingBlock
+                                let cont = ExecuteBlock $ runMVR exec mvr
+                                return (updateResult, Just cont)
+        (EVersionedConfigurationV1 (vc :: VersionedConfigurationV1 finconf pv)) -> do
+            MVR $ \mvr -> do
+                now <- currentTime
+                case SkovV1.deserializeExactVersionedPendingBlock @pv blockBS now of
+                    Left err -> do
+                        mvLog mvr Runner LLDebug err
+                        return (Skov.ResultSerializationFail, Nothing)
+                    Right block -> do
+                        blockResult <-
+                            runMVR
+                                (runSkovV1Transaction vc (SkovV1.uponReceivingBlock block))
+                                mvr
+                        case blockResult of
+                            SkovV1.BlockResultSuccess vb -> do
+                                let exec = do
+                                        runSkovV1Transaction vc (SkovV1.executeBlock vb)
+                                        return Skov.ResultSuccess
+                                let cont = ExecuteBlock $ runMVR exec mvr
+                                return (Skov.ResultSuccess, Just cont)
+                            SkovV1.BlockResultDoubleSign vb -> do
+                                runMVR (runSkovV1Transaction vc (SkovV1.executeBlock vb)) mvr
+                                return (Skov.ResultDoubleSign, Nothing)
+                            SkovV1.BlockResultInvalid -> return (Skov.ResultInvalid, Nothing)
+                            SkovV1.BlockResultStale -> return (Skov.ResultStale, Nothing)
+                            SkovV1.BlockResultPending -> return (Skov.ResultPendingBlock, Nothing)
+                            SkovV1.BlockResultEarly -> return (Skov.ResultEarlyBlock, Nothing)
+                            SkovV1.BlockResultDuplicate -> return (Skov.ResultDuplicate, Nothing)
+                            SkovV1.BlockResultConsensusShutdown ->
+                                return (Skov.ResultConsensusShutDown, Nothing)
 
 -- | Invoke the continuation yielded by 'receiveBlock'.
 --  The continuation performs a transaction which will acquire the write lock
 --  before trying to add the block to the tree and release the lock again afterwards.
 executeBlock :: ExecuteBlock -> MVR finconf Skov.UpdateResult
-executeBlock = liftIO . runBlock
+executeBlock = handleMVRExceptions . liftIO . runBlock
 
 -- | Deserialize and receive a finalization message at a given genesis index.
 receiveFinalizationMessage :: GenesisIndex -> ByteString -> MVR finconf Skov.UpdateResult
-receiveFinalizationMessage gi finMsgBS = withLatestExpectedVersion_ gi $ \case
+receiveFinalizationMessage gi finMsgBS = handleMVRExceptions $ withLatestExpectedVersion_ gi $ \case
     (EVersionedConfigurationV0 (vc :: VersionedConfigurationV0 finconf pv)) ->
         case runGet getExactVersionedFPM finMsgBS of
             Left err -> do
@@ -1748,16 +1817,16 @@ receiveFinalizationMessage gi finMsgBS = withLatestExpectedVersion_ gi $ \case
                             Left leftRes -> (leftRes, Nothing)
                             Right cont -> (Skov.ResultSuccess, Just cont)
                     followup = liftSkovV1Update vc
-                -- We spawn a thread to perform the follow-up so that the P2P layer can immediately
-                -- relay the message, since the follow-up action can be time consuming (including
-                -- finalizing blocks and baking a new block).
+                -- We spawn a thread to perform the follow-up so that the P2P layer can
+                -- immediately relay the message, since the follow-up action can be time
+                -- consuming (including finalizing blocks and baking a new block).
                 withWriteLockMaybeFork receive followup
 
 -- | Deserialize and receive a finalization entity.
 --  For consensus version 0 this should be a 'FinalizationRecord'.
 --  For consensus version 1 this should be a 'FinalizationEntry'.
 receiveFinalization :: GenesisIndex -> ByteString -> MVR finconf Skov.UpdateResult
-receiveFinalization gi finBS = withLatestExpectedVersion_ gi $ \case
+receiveFinalization gi finBS = handleMVRExceptions $ withLatestExpectedVersion_ gi $ \case
     (EVersionedConfigurationV0 (vc :: VersionedConfigurationV0 finconf pv)) ->
         case runGet getExactVersionedFinalizationRecord finBS of
             Left err -> do
@@ -1794,7 +1863,7 @@ receiveCatchUpStatus ::
     CatchUpConfiguration ->
     MVR finconf Skov.UpdateResult
 receiveCatchUpStatus gi catchUpBS cuConfig@CatchUpConfiguration{..} =
-    case runGet getVersionedCatchUpStatus catchUpBS of
+    handleMVRExceptions $ case runGet getVersionedCatchUpStatus catchUpBS of
         Left err -> do
             logEvent Runner LLDebug $ "Could not deserialize catch-up status message: " ++ err
             return Skov.ResultSerializationFail
@@ -1988,7 +2057,7 @@ getCatchUpRequest = do
 --  the result of the update. The hash is present unless the transaction could
 --  not be deserialized.
 receiveTransaction :: forall finconf. ByteString -> MVR finconf (Maybe TransactionHash, Skov.UpdateResult)
-receiveTransaction transactionBS = do
+receiveTransaction transactionBS = handleMVRExceptionsWith (Nothing, Skov.ResultConsensusFailure) $ do
     now <- utcTimeToTransactionTime <$> currentTime
     mvr <- ask
     vvec <- liftIO $ readIORef $ mvVersions mvr
@@ -2094,7 +2163,7 @@ receiveExecuteBlock gi blockBS = withLatestExpectedVersion_ gi $ \case
 
 -- | Import a block file for out-of-band catch-up.
 importBlocks :: FilePath -> MVR finconf Skov.UpdateResult
-importBlocks importFile = do
+importBlocks importFile = handleMVRExceptions $ do
     vvec <- liftIO . readIORef =<< asks mvVersions
     -- Import starting from the genesis index of the latest consensus
     let genIndex = evcIndex (Vec.last vvec)